Move Thumbs component to lib
