+++ client/src/lib/components/Thumbs/Thumbs.tsx
